requirements.txt
ops>=1.3.0,<2.0.0
lightkube>=0.10.1,<1.0.0
jinja2
jsonschema
pyyaml